+++ manager-ui/i18n/locales/en_GB/app.ts
@@ -1,6 +1,5 @@
 import type { Locales } from "../../Locales";
 
 export default {
-  name: "Simply Mail Server",
   language: "Language",
 } satisfies Locales["app"];